Quick simple tests for podman
This is a placeholder only. The idea is that, once the ginkgo
tests are approved and merged and packaged, you add that
package to test_podman.yml:required_packages, then rewrite
the .sh script so it invokes those tests and writes PASS/FAIL
to test.log as appropriate.

Until then, this can serve as a rudimentary gate.

Signed-off-by: Ed Santiago <santiago@redhat.com>