gating: change the PassingTestCaseRule string
The first document I went with[1] specified a magic
value containing 'package.test.functional'. I used
that, and my bodhi results[2] looked weird: first,
none of the usual results include 'test.functional';
so, second, there's now a new test.functional line
under 'undefined' in red and with an asterisk next
to it, and another red one without an asterisk.
All four lines seem to link to the same results.

 [1] https://docs.fedoraproject.org/en-US/ci/gating/
 [2] https://bodhi.fedoraproject.org/updates/FEDORA-2019-144ed359e9

A different document[3] specifies a magic value that
more closely conforms to the ones already shown
in the 'undefined' section of older bodhi runs.
That's what I'm trying now.

 [3] https://docs.fedoraproject.org/en-US/rawhide-gating/optin/

Unfortunately neither document has a timestamp so it's
impossible to know which (if any) is up-to-date and/or
correct.

Signed-off-by: Ed Santiago <santiago@redhat.com>