Sigh. Put the reboot in a block.
Podman gating tests failed:

   ... back_again.changed): 'dict object' has no attribute 'changed'

Previous error (ignored) was a timeout in the reboot step. I have
no idea how to fix this, so let's try putting the reboot and wait
steps into a 'block' section and removing the '.changed'.

Tested with 1minutetip. Seems to work, but the reboot was quick
so it didn't trigger the same code.

Signed-off-by: Ed Santiago <santiago@redhat.com>