Sigh. Disable reboot. Disable cgroups v1 test.
There's a nasty bug in the CI infrastructure: reboots
hang forever[1]. Sometimes they work, but only 25% of
the time -- the rest of the time, gating tests fail.

 [1] https://redhat.service-now.com/surl.do?n=PNT0808530

So, let's just disable the reboot and all cgroups v1 tests.
For now, I've disabled with a quick edit that should allow
us to reenable if/when the reboot bug is fixed.

TODO: reevaluate in three months (2020-08). If the reboot
bug is not fixed by then, just scrap the entire cgroups
refactoring and go back to plain standard-test-basic.

Signed-off-by: Ed Santiago <santiago@redhat.com>